anisotropyChoice(intamapInteractive)	R Documentation

anisotropyChoice

Description

This function combines segmentation of scattered 2D data and estimation of anisotropy parameters using the CTI method.

Usage

anisotropyChoice(object)

Arguments

object 	 An Intamap type object containing one SpatialPointsDataFrame with observations.

Details

The function AnisotropyChoice function employs the doSegmentation function to automatically separate the original dataset into clusters based on the sampling density and the spatial locations of the data (see doSegmentation for details). The results of the segmentation procedure and the anisotropy analysis per cluster are returned in a matrix of dimension [cl]x5, where [cl] is the number of clusters . Each row of the matrix contains the cluster index, the anisotropy ratio, the anisotropy direction, the number of cluster points and the area inside the convex hull of the cluster. In addition, a single set of anisotropy parameters is returned in the element anisPar. These parameters are calculated using weighted averages of the covariance Hessian matrix estimates in each cluster. The weights are based on the area enclosed by the convex hull of each cluster.

Value

object: A modified Intamap type object is returned, which contains the results of the anisotropy parameter estimation. The anisotropy parameters are returned in the element anisPar as described below.

anisPar 	 List element in object that contains a list with the following elements:

ratio

A coarse-grained anisotropy ratio for all the data

direction

A coarse-grained anisotropy orientation for all the data

clusters

A matrix of dimension [cl]x5 which determines the anisotropy per cluster. Each row of clusters gives the (cluster id, anisotropy ratio, anisotropy direction, number of points, area) for each cluster detected.

clusters 	 list element added to the original object containing the segmentation results.

index

Index array identifying the cluster in which each observation point belongs. Zero value means that the observations has been removed.

clusterNumber

Number of clusters detected.

Note

This function uses the akima package to perform "bilinear" and "bicubic" interpolation for the estimation of spatial derivatives

Author(s)

D.T. Hristopulos, G.Spiliopoulos, A.Chorti

References

[1] http://www.intamap.org

[2] A. Chorti and D. T. Hristopulos (2008). Non-parametric Identification of Anisotropic (Elliptic) Correlations in Spatially Distributed Data Sets, IEEE Transactions on Signal Processing, 56(10), 4738-4751 (2008).

[3] D. T. Hristopulos, M. P. Petrakis, G. Spiliopoulos, A. Chorti (2009). Non-parametric estimation of geometric anisotropy from environmental sensor network measurements, StatGIS 2009: Geoinformatics for Environmental Surveillance Proceedings (ed. G. Dubois).

Examples

library(intamapInteractive)

data(walker)
coordinates(walker)=~X+Y
object=createIntamapObject(observations=walker)
object=anisotropyChoice(object)

print(summary(object$clusters$index))
print(object$anisPar)

[Package intamapInteractive version 1.0-6 Index]