calculateMukv(intamapInteractive)	R Documentation

MUKV computation

Description

Computes MUKV for given geostatistical parameters

Usage

calculateMukv(observations, predGrid, model, formulaString, ...)

Arguments

observations 	 Data frame with x,y coordinates of observations and possible covariates

predGrid 	 Data frame with x,y coordinates of prediction locations

model 	 Variogram model:object of class variogramModel, of the form created by vgm

formulaString 	 formula that defines the dependent variable as a linear model of independent variables; suppose the dependent variable has name z, for ordinary and simple kriging use the formula z~1; for universal kriging, suppose z is linearly dependent on x and y, use the formula z~x+y. The formulaString defaults to "value~1" if value is a part of the data set. If not, the first column of the data set is used.

... 	 other arguments to be passed on at lower level functions

Details

This function computes kriging on the predGrid with krige function, and averages the kriging variance (MUKV). With covariates, the function takes a universal kriging model into account.

Value

MUKV value

Author(s)

S.J. Melles and O. Baume

References

http://www.intamap.org/

[Package intamapInteractive version 1.0-6 Index]