checkSetup(intamap)	R Documentation

check setup

Description

checkSetup will do some sanity checks on input data provided through object.

Usage

checkSetup(object, quiet = FALSE)

Arguments

object 	 object, to be passed to preProcess, see intamap-package

quiet 	 logical; TRUE to suppress OK statement

Details

checkSetup is a function that makes certain tests on the intamap object to make sure that it is suited for interpolation. Particularly, it will issue a warning or an error if one of the following conditions are met:

observations is not an element of object
observations contain less than 20 observations
Some of the observation locations are duplicated
formulaString is not an element of object
None of the columns of observations has a name that corresponds to the independent variable of formulaString
predictionLocations is not an element of object
predictionLocations is not a Spatial object
targetCRS is given, but observations and predictionLocations do not have CRS set
addBias includes biases that are not part of removeBias

The function will issue a warning if it appears that predictionLocations and observations share a small region. This warning is given as it is a likely cause of errors, although it can also happen if predictionLocations are limited to one small cluster.

Value

returns TRUE if check passes, will halt with error when some some error condition is met.

Author(s)

Edzer J. Pebesma

References

http://www.intamap.org/

[Package intamap version 1.2-18 Index]