copulaEstimation(intamap)	R Documentation

ML-estimation of the spatial copula model parameters

Description

Estimates parameters of the spatial copula model using maximum likelihood.

Usage

copulaEstimation(obj,margin,trend,correlation,anisotropy,copula,tol=0.001,...)

Arguments

obj 	 Intamap object, see description in intamap-package

margin 	 list with the following elements:

params

Starting values for the parameters of the marginal distribution (excluding trend parameters)

lower

Lower bounds for the values of the parameters of the marginal distribution (excluding trend parameters)

upper

Upper bounds for the values of the parameters of the marginal distribution (excluding trend parameters)

name

Name of the family of marginal distributions. Possible names are: "norm","lnorm","gev","t" and "logis"

trend 	 list with the following elements:

params

Starting values for the parameters of the trend model (location parameter of the marginal distribution)

lower

Lower bounds for the values of the parameters of the trend model

upper

Upper bounds for the values of the parameters of the trend model

F

Design matrix.

correlation 	 list with the following elements:

model

Correlation function model. Possible models are: "Ste", "Sph", "Gau" and "Exp"

params

Starting values for the parameters of the correlation function model

lower

Lower bounds for the values of the parameters of the correlation function model

upper

Upper bounds for the values of the parameters of the correlation function model

anisotropy 	 list with the following elements:

params

Starting values for the parameters of geometric anisotropy. If NULL, then no anisotropy is considered.

lower

Lower bounds for the values of the parameters of geometric anisotropy. Usually c(0,1)

upper

Upper bounds for the values of the parameters of geometric anisotropy. Usually c(pi,Inf)

copula 	 list with the following elements:

method

Either "norm" or "chisq", depending on which spatial copula model is used, the Gaussian or the chi-squared copula.

params

Only used in case of the chi-squared copula: the squared non-centrality parameter of the non-central chi-squared distribution. Controls how far the chi-squared copula is from the Gaussian copula.

lower

Only used in case of the chi-squared copula: the lower bound for the copula parameter. Usually set to 0

upper

Only used in case of the chi-squared copula: the upper bound for the copula parameter. Usually set to Inf

tol 	 Tolerance level for the optimization process.

... 	 Arguments to be passed to optim.

Details

copulaEstimation performs maximum likelihood estimation of all possible parameters included in the Gaussian and chi-squared spatial copula model: parameters of the predefined family of marginal distributions (including spatial trend or external drift), correlation function parameters, parameters for geometric anisotropy and parameters for the copula (only used for the chi-squared copula model). Due to the large number of variables that need to be optimized, a profile-likelihood approach is used. Although convergence to a global optimum is not assured, the profile-likelihood method makes it less likely that the optimization routine, optim, gets stuck in a local optimum. The result of copulaEstimation is a list containing all parameter point estimates that are needed for plug-in spatial prediction. It is advisable to check the output of the algorithm by trying different starting values for the optimization.

Value

A list with the following elements:

margin 	 Same as the input except that the list element "params" now consists of the optimized parameters of the marginal distribution function.

trend 	 Same as the input except that the list element "params" now consists of the optimized parameters of the trend model.

correlation 	 Same as the input except that the list element "params" now consists of the optimized parameters of the correlation function model.

anisotropy 	 Same as the input except that the list element "params" now consists of the optimized parameters of geometric anisotropy.

copula 	 Same as the input except that the list element "params" now consists of the optimized copula parameters.

Author(s)

Hannes Kazianka

References

Kazianka, H. and Pilz, J. (2009), Spatial Interpolation Using Copula-Based Geostatistical Models. GeoENV2008 - Geostatistics for Environmental Application (P. Atkinson, C. Lloyd, eds.), Springer, New York

See Also

bayesCopula, spatialPredict, estimateParameters

Examples

data(intamapExampleObject)
## estimate parameters for the copula model

## Not run: 
copula<-list(method="norm")
anisotropy<-list(lower=c(0,1),upper=c(pi,Inf),params=c(pi/3,2))
correlation<-list(model="Ste",lower=c(0.01,0.01,0.01),upper=c(0.99,Inf,20),params=c(0.05,4,3))
margin<-list(name="gev",lower=c(0.01,-Inf),upper=c(Inf,Inf),params=c(30,0.5))
trend<-list(F=as.matrix(rep(1,196)),lower=-Inf,upper=Inf,params=40)
estimates<-copulaEstimation(intamapExampleObject,margin,trend,correlation,anisotropy,copula)
## make predictions at unobserved locations
predictions<-bayescopula(intamapExampleObject,estimates,search=25,calc=list(mean=TRUE,variance=TRUE,excprob=40,quantile=0.95))
## End(Not run)

[Package intamap version 1.2-18 Index]