estimateTimeModel(intamap)	R Documentation

estimateTimeModel

Description

Function that takes time samples function that can read intamap objects

Usage

estimateTimeModel(FUN, class, formulaString, debug.level, ...)

Arguments

FUN 	 A string with function's name

class 	 class of intamapObject, which interpolation method to be used.

formulaString 	 the formula of the request, mainly to see if the request has independent variables.

debug.level 	 if debug.level >= 1, the function will store tables with the prediction times for each model in the workspace.

... 	 other arguments as defined in the createIntamapObject function

Details

This function uses createIntamapObject function to create synthetic data, in order to take time samples for the function with string name "FUN". The calculated model is stored, as an element of a list, in a local file (workspace for now) and it's used in order to give quick time estimates.

Value

The function does not return a variable but stores the result in an element list with the same name.

[Package intamap version 1.2-13 Index]