findBiasUK(intamapInteractive)	R Documentation

Finding the regional biases using GLM

Description

Method for identifying regional biases (in most cases biases between countries)

Usage

findBiasUK(object)

Arguments

object 	 an object of class SpatialPointsDataFrame, at least containing observations and a regional identification code (regCode)

Value

A data.frame with the biases for each country with uncertainty.

Author(s)

Olivier Baume

References

http://www.intamap.org/

See Also

findRegionalBias

[Package intamapInteractive version 1.0-6 Index]