generateTimeModels(intamap)	R Documentation

Generate time models

Description

function that generates time models and saves them in workspace.

Usage

generateTimeModels(genClasses = NULL, noGenClasses = NULL, nSam = 1, test = FALSE,
             debug.level = 0)

Arguments

genClasses 	 list of particular classes for which time models should be generated

noGenClasses 	 list of particular classes for which time models should not be generated

nSam 	 number of attempts to be tried for each combination of predictions and observations, defaults to 1, higher number should be used for better accuracy. nSam/2 is used for copulas, to reduce computation time.

test 	 logical; if true, the time models are generated based on fewer iterations, for speed

debug.level 	 if debug.level >= 1, the function will store tables with the prediction times for each model in the workspace.

Details

This function calculates a time model for different interpolation types in the intamap-package and returns a list object with the estimated models. It's users responsibility to store the model in the workspace. The normal procedure would be to run the function without arguments. However, it is both possible to define a list for which classes the user want to generate models, or a list of classes that are not of interest.

The time model is based on creation of a set of synthetical data sets of different size, both regarding number of observations and prediction locations. The function will estimate parameters and make predictions with the different combinations, and for each method, fit a local polynomial regression model (loess)

This model can then be used by predictTime to estimate the prediction time for an interpolation request with a certain number of observations and prediction locations.

Value

The function generates a timeModels object, which can be used to estimate prediction times for different requests to the interpolate function in the intamap-package, via predictTime.

Examples

## Not run: 
timeModels=generateTimeModels()
q("yes")
## restart R in the same directory
## End(Not run)

[Package intamap version 1.2-18 Index]