getIntamapParams(intamap)	R Documentation

Setting parameters for the intamap package

Description

This function sets a range of the parameters for the intamap-package, to be included in the object described in intamap-package

Usage

getIntamapParams(oldPar,newPar,...)

Arguments

oldPar 	 An existing set of parameters for the interpolation process, of class IntamapParams or a list of parameters for modification of the default parameters

newPar 	 A list of parameters for updating oldPar or for modification of the default parameters. Possible parameters with their defaults are given below

... 	 Individual parameters for updating oldPar or for modification of the default parameters. Possible parameters with their defaults are given below

doAnisotropy = FALSE

Defining whether anisotropy should be calculated

removeBias = NA

Defining whether biases should be removed, and in case yes, which ones (localBias and regionalBias implemented

addBias = NA

Defining which biases to be added in the postProcess function. This has not yet been implemented.

biasRemovalMethod = "LM"

character; specifies which methods to use to remove bias. See below.

doSegmentation = FALSE

Defining if the predictions should be subject to segmentation. Segmentation has been implemented, but not the use of it.

testMean

logical; for copula method only; whether or not the predictive means (if calculated) should be tested for being reasonable

nmax = 50

for local kriging: the number of nearest observations that should be used for a kriging prediction or simulation, where nearest is defined in terms of the space of the spatial locations. By default, 50 observations are used.

ngrid = 100

The number of grid points to be used if an Averaged Cumulative Distribution Function (ACDF) needs to be computed for unbiased kriging

nsim=100

Number of simulations when needed

block = numeric(0)

Block size; a vector with 1, 2 or 3 values containing the size of a rectangular in x-, y- and z-dimension respectively (0 if not set), or a data frame with 1, 2 or 3 columns, containing the points that discretize the block in the x-, y- and z-dimension to define irregular blocks relative to (0,0) or (0,0,0) - see also the details section of predict.gstat. By default, predictions or simulations refer to the support of the data values.

processType = "gaussian"

If known - the distribution of the data. Defaults to gaussian, analytical solutions also exists in some cases for logNormal. This setting only affects a limited number of methods, e.g. the block prediciton

confProj = FALSE

If set, the program will attempt conform projections in preProcess, calling the function conformProjections.

debug.level = 0

Used in some functions for giving additional output. See individual functions for more information.

...

Additional parameters that do not exist in the default parameter set, this could be parameters necessary for new methods within or outside the intamap-package

Value

A list of the parameters with class intamapParams to be included in the object described in intamap-package

Note

This function will mainly be called by createIntamapObject, but can also be called by the user to create a parameter set or update an existing parameter set. If none of the arguments is a list of class IntamapParams), the function will assume that the argument(s) are modifications to the default set of parameters.

If the function is called with two lists of parameters (but the first one is not of class IntamapParams) they are both seen as modifications to the default parameter set. If they share some parameters, the parameter values from the second list will be applied.

Author(s)

Jon Olav Skoien

References

http://www.intamap.org/

See Also

createIntamapObject

Examples

# Create a new set of intamapParameters, with default parameters:
params = getIntamapParams()
# Make modifications to the default list of parameters
params = getIntamapParams(newPar=list(removeBias = "local",
              secondParameter = "second"))
# Make modifications to an existing list of parameters
params = getIntamapParams(oldPar = params,newPar = list(predictType = list(exc=TRUE)))

[Package intamap version 1.2-13 Index]