intamapExampleObject(intamap)	R Documentation

Simulated Intamap Object

Description

Intamap object of class "copula" containing a simulated data set with 196 spatial locations.

Usage

data(intamapExampleObject)

Details

The data set is a realization of a random field generated using a Gaussian copula and generalized extreme value distributed margins (location=40,shape=0.5, scale=30). The correlation function is Matern (Stein's representation) with range=4, kappa=3 and nugget=0.05. Furthermore, there is geometric anisotropy with direction=pi/3 and ratio=2.

See Also

spatialPredict.copula, estimateParameters.copula

Examples

## Not run: 
data(intamapExampleObject)
## estimate parameters for the copula model
intamapExampleObject<-estimateParameters(intamapExampleObject)
## make predictions at unobserved locations
intamapExampleObject<-spatialCopula(intamapExampleObject)
## End(Not run)

[Package intamap version 1.2-12 Index]