intamapInteractive-package(intamapInteractive)	R Documentation

Interactive functionality added to the intamap package

Description

This package provides some added functionality to the link[intamap]{intamap-package} for automatic interpolation of environmental variables. Whereas link[intamap]{intamap-package} was particularly developed as a statistical back-end of a Web Processing Service (WPS), does this package offer some functionality that is not possible to access through such a WPS.

The methods in this package can mainly be put into three groups:

bias correction

methods for estimating and possible correct for biases between measurement networks, due to differences in measurement strategies, measurement devices, or (unknown) post-processing of data

segmentation

method for segmentation of data, based on their measurement density

network optimization

methods for optimizing a measurement network (adding or removing observation points), based on different criteria

[Package intamapInteractive version 1.0-6 Index]