methodParameters(intamap)	R Documentation

generate string for generation of method parameters

Description

function that generates a parsable string of identified method parameters for an intamap interpolation object

Usage

## Default S3 method:
methodParameters(object)
## S3 method for class 'copula':
methodParameters(object)
## S3 method for class 'idw':
methodParameters(object)

Arguments

object 	 a list object. Most arguments necessary for interpolation are passed through this object. See intamap-package for further description of the necessary content of this variable

Details

The function creates a text-string that makes it possible to add the the method parameters (anisotropy and idw-parameter, variogram model or copula parameters) to the object in a later call to createIntamapObject or interpolate without having to re-estimate the parameters. This function is particularly useful when interpolate is called from a Web Processing Service, and the user wants to reuse the method parameters. The function is mainly assumed to be called from within interpolate.

The default method assumes a variogram model of gstat type, e.g. a variogram similar to what can be created with a call to vgm. Also psgp uses this variogram model.

Value

A string that, when parsed, will recreate the methodParameters

Author(s)

Jon Olav Skoien

References

http://www.intamap.org/

Examples

library(intamap)
data(meuse)
coordinates(meuse) = ~x+y
meuse$value = log(meuse$zinc)
# set up intamap object:
krigingObject = createIntamapObject(
                observations = meuse,
                formulaString = as.formula('value~1'),class = "automap")
# do estimation steps:
krigingObject = estimateParameters(krigingObject)
krigingObject = methodParameters(krigingObject)

# Create a new object
krigingObject2 = createIntamapObject(observations = meuse,
                formulaString = as.formula('value~1'),
    params = list(methodParameters = krigingObject$methodParameters))

krigingObject$variogramModel
krigingObject2$variogramModel

[Package intamap version 1.2-18 Index]