plotIntamap(intamap)	R Documentation

plot intamap objects

Description

Plotting function for intamap-objects of the type described in intamap-package

Usage

plotIntamap(object, zcol = "all", sp.layout = NULL, plotMat = c(2,2), ...)
## S3 method for class 'copula':
plot(x,  ...)
## S3 method for class 'idw':
plot(x, ...)
## S3 method for class 'automap':
plot(x, ...)
## S3 method for class 'linearVariogram':
plot(x, ...)
## S3 method for class 'transGaussian':
plot(x, ...)
## S3 method for class 'yamamoto':
plot(x, ...)

Arguments

object 	 a list object. Most arguments necessary for interpolation are passed through this object. See intamap-package for further description of the necessary content of this variable

x 	 intamap object, when plot is called directly

zcol 	 a list of column names to be plotted; if equal to all, the column names will correspond to all possible column names from outputWhat (see createIntamapObject)

sp.layout 	 an object that can contain lines, points and polygons that function as extra layout; see spplot for more information

plotMat 	 an array of lengt two with the number of rows and columns of plots per page

... 	 other parameters that can be passed to other plot functions (e.g. plot, spplot, automapPlot and xyplot

Details

All the plot methods above are simple wrapper functions around the plotIntamap function.

Value

A plot of some of the elements of object. This will typically be the sample variogram and the fitted variogram model (if a method based on variograms has been used) and all the predictions.

Author(s)

Jon Olav Skoien

References

http://www.intamap.org/

Examples

data(meuse)
meuse$value = log(meuse$zinc)
data(meuse.grid)
coordinates(meuse) = ~x+y
coordinates(meuse.grid) = ~x+y
object = interpolate(meuse,meuse.grid,outputWhat = list(mean = TRUE, variance = TRUE, excprob = 7, excprob = 8, quantile = 0.9, quantile = 0.95),methodName = "automap")
plot(object)

[Package intamap version 1.2-18 Index]