postProcess(intamap)	R Documentation

pre-process data

Description

post-processing of data for the intamap-package. The function will typically call other functions for adding back biases, aggregation etc.

Usage

## Default S3 method:
postProcess(object, ...)
## S3 method for class 'idw':
postProcess(object, ...)

Arguments

object 	 a list object. Most arguments necessary for interpolation are passed through this object. See intamap-package for further description of the necessary content of this variable

... 	 other parameters that can be passed to functions called from preProcess

Details

The function postProcess includes code for postprocessing an object after interpolation. The function can easily be replaced by more specific methods relevant for a certain data set, doing more data specific things in addition to what is done in the default method.

Value

An object of same type as above, but with new elements. Most important from the default method is the outputTable, a matrix, organized in a convenient way for the calling WPS; first row: x-coordinates, second row: y-coordinates; further rows: output elements as specified by outputWhat (see createIntamapObject

Author(s)

Edzer J. Pebesma

References

http://www.intamap.org/

[Package intamap version 1.2-18 Index]