preProcess(intamap)	R Documentation

pre-processing of data

Description

pre-processing of the data for the intamap-package package.

Usage

## Default S3 method:
preProcess(object, ...)
## S3 method for class 'idw':
preProcess(object, ...)

Arguments

object 	 see intamap-package; list that should at least contain (i) an element called observations of class SpatialPointsDataFrame. The measured values should be named value, and (ii) an element params of class list, by calling the function getIntamapParams. (iii) Usually, the object will also contain an element called predictionLocations, of a class extending Spatial.

... 	 Additional parameters

Details

The function preProcess includes code for preprocessing an object before interpolation. The function can easily be replaced by more specific methods relevant for a certain data set. Functions can be called from preProcess according to the settings in parameters in the object, set by the function getIntamapParams.

Value

The input object is returned, after its components have been pre-processed.

Author(s)

Jon Olav Skoien

References

http://www.intamap.org/

[Package intamap version 1.2-18 Index]