predictTime(intamap)	R Documentation

Time prediction for intamap package methods

Description

Functions that gives a time estimate for an intamap function given the number of observations and predictionLocations

Usage

predictTime(nObs, nPred, class, formulaString, calibration=FALSE, 
                   outputWhat, FUN = "spatialPredict",...)

Arguments

nObs 	 An integer or an array of integers containing the number of observations.

nPred 	 An integer or an array of integers containing the number of predictions.

class 	 class of intamapObject, which interpolation method to be used

formulaString 	 the formula of the request, mainly to see if the request has independent variables

calibration 	 enables or disables time calibration - not properly implemented yet

outputWhat 	 List defining the requested type of output, see createIntamapObject

FUN 	 A string with the intamap functions name, now obsolete

... 	 other arguments needed to define the intamap object.

Details

The function is based on timeModels being available in the workspace. This is a loess-model that has been fitted to different calls to a range of interpolation requests with synthetically generated data in generateTimeModels.

Value

An integer or an array of integers with the predicted times.

Note

RUN FIRST generateTimeModels() or estimateTimeModel() in order to save time Models to workspace. It might take some time!

[Package intamap version 1.2-18 Index]