rotateAnisotropicData(intamap)	R Documentation

rotateAnisotropicData

Description

This function applies an isotropic transformation of the coordinates specified in object.

Usage

rotateAnisotropicData(object,anisPar)

Arguments

object 	 (i) An Intamap type object (see intamap-package) containing one SpatialPointsDataFrame data frame named observations which includes the observed values (ii) or a SpatialPointsDataFrame which includes both coordinates and observations or (iii) SpatialPoints which includes only coordinates to be rotated.

anisPar 	 An array containing the anisotropy parameters (anisotropy ratio and axes orientation) (see estimateAnisotropy) for the rotation. If object is the output of estimateAnisotropy function, these parameters are part of object. In cases (ii) and (iii) anisPar defines the two anisotropy parameters. For the definition of the anisotropy parameters see estimateAnisotropy.

Details

This function performs a rotation and rescaling of the coordinate axes in order to obtain a new coordinate system, in which the observations become statistically isotropic. This assumes that the estimates of the anisotropy ratio and the orientation angle provided in anisPar are accurate.

Value

(i) A modified object with transformed coordinates if rotateAnisotropicData is called with an Intamap object as input (see intamap-package) or (ii) the transformed coordinates if a SpatialPointsDataFrame is used as input or (iii) the transformed coordinates if a SpatialPoints object is the input.

Author(s)

Hristopulos Dionisis, Spiliopoulos Giannis

References

[1] http://www.intamap.org

[2] A. Chorti and D. T. Hristopulos (2008). Non-parametric Identification of Anisotropic (Elliptic) Correlations in Spatially Distributed Data Sets, IEEE Transactions on Signal Processing, 56(10), 4738-4751 (2008).

See Also

estimateAnisotropy

Examples

    library(gstat)
    data(sic2004)
    coordinates(sic.val)=~x+y
    sic.val$value=sic.val$dayx

    params=NULL
    
    obj<-list(
        observations=sic.val
        )
    obj<-estimateAnisotropy(obj)
    print(obj$anisPar)  

    obj$observations<-rotateAnisotropicData(obj$observations,obj$anisPar)

    obj<-estimateAnisotropy(obj)
    print(obj$anisPar)  

[Package intamap version 1.2-18 Index]