spCovAdd(intamapInteractive)	R Documentation

Spatial coverage method to add new measurements

Description

This function spCovAdd allows to build optimization scenarios based on spatial coverage method.

Usage

spCovAdd( observations, candidates, nDiff, nGridCells, plotOptim = TRUE, nTry, ... )

Arguments

observations 	 object of class data.frame with x,y coordinates

candidates 	 a SpatialPolygonsDataFrame to explore: in use when optimizing the implementation of new measurement stations to an existing network

nDiff 	 number of stations to add or delete

nGridCells 	 number of grid cells to work on spatial coverage strafication

plotOptim 	 logical; to plot the result or not

nTry 	 the method will try nTry initial configurations and will keep the best solution in order to reduce the risk of ending up with an unfavorable solution

... 	 other arguments to be passed on at lower level functions such as stratify

Details

This function allows to build optimization scenarios based on spatial coverage method. The scenario action is "add". To add new measurement locations to the running network, the function uses function stratify from package spcosa. Function stratify adds new strata to the domain study.

Value

data.frame of optimized locations

Author(s)

Olivier Baume

References

[1] http://www.intamap.org/

[2] D. J. Brus, J. de Gruijter, J. van Groenigen (2006). Designing spatial coverage samples using the k-means clustering algorithm. In A. McBratney M. Voltz and P. Lagacherie, editor, Digital Soil Mapping: An Introductory Perspective, Developments in Soil Science, vol. 3., Elsevier, Amsterdam.

[Package intamapInteractive version 1.0-6 Index]