spCovDel(intamapInteractive)	R Documentation

Optimize the network with spatial coverage methods

Description

The function spCovDel allows to build optimization scenarios based on spatial coverage method.

Usage

spCovDel(observations, candidates, nDiff, plotOptim = TRUE, ...)

Arguments

observations 	 object of class data.frame with x,y coordinates

candidates 	 not compulsory used only for plotting purpose – a SpatialPolygonsDataFrame describing the study area

nDiff 	 number of stations to add or delete

plotOptim 	 logical; to plot the result or not

... 	 other arguments to be passed on at lower level functions such as nndist

Details

This function allows to build optimization scenarios based on spatial coverage method. When action is "del", the function maximizes the mean distance of measurements with direct neighbours using function nndist from package spatstat. The heuristic search uses a swapping algorithm to converge more rapidly to the best solution.

Value

data.frame of optimized locations

Author(s)

Olivier Baume

References

http://www.intamap.org/

[Package intamapInteractive version 1.0-6 Index]