ssaOptim(intamapInteractive)	R Documentation

Spatial simulated annealing for optimized model based mapping

Description

nDiff measurement stations will be optimally Optimization of measurement network with Spatial simulated annealing and MUKV

Usage

ssaOptim(observations, predGrid, candidates, action, nDiff, model,
nr_iterations, plotOptim, ...)

Arguments

observations 	 Data frame with x,y coordinates of observations and possible covariates

predGrid 	 Data frame with x,y coordinates of prediction locations when method is "ssa"

candidates 	 When action is "add", candidates is data frame with x,y coordinates or a SpatialPolygonsDataFrame

action 	 Character string indicating which type of action to simulate: "add" to add new measurement stations to the existing network or "del" to turn off existing stations

nDiff 	 Number of stations to add or delete

model 	 Variogram model:object of class variogramModel, as generated by vgm

nr_iterations 	 number of iterations before stoping

plotOptim 	 plot or not the result; TRUE by default

... 	 other arguments to be passed on lower level functions

Details

This function applies SSA algorithm for the special case of optimization criterion MUKV. With covariates, the function takes a universal kriging model into account. Stopping criterion of the optimization procedure uses a number of searching steps without improvement in the design.

Value

Data frame of optimized locations

Author(s)

Olivier Baume

References

http://www.intamap.org/

[Package intamapInteractive version 1.0-6 Index]