summaryIntamap(intamap)	R Documentation

summary intamap objects

Description

summary function for intamap-objects of the type described in intamap-package

Usage

summaryIntamap(object, ...)
## S3 method for class 'copula':
summary(object, ...)
## S3 method for class 'idw':
summary(object, ...)
## S3 method for class 'automap':
summary(object, ...)
## S3 method for class 'linearVariogram':
summary(object, ...)
## S3 method for class 'transGaussian':
summary(object, ...)
## S3 method for class 'yamamoto':
summary(object, ...)

Arguments

object 	 a list object. Most arguments necessary for interpolation are passed through this object. See intamap-package for further description of the necessary content of this variable

... 	 parameters to be passed to the default summary function for each element

Value

A summary of some of the elements of object.

Author(s)

Jon Olav Skoien

References

http://www.intamap.org/

[Package intamap version 1.2-18 Index]