timeModels(intamap)	R Documentation

models for estimating prediction time in intamap package

Description

This is a standard model for estimating the prediction time within the intamap-package. It was created by the function generateTimeModels, on a 64 bits Linux server running R version 2.9.0 and intamap-package version 1.1.15.

The prediction time will depend on the speed of the local machine, on the version of R and intamap-package, and on the installed libraries. It is therefore strongly recommended to run generateTimeModels on the local machine and store the result in the workspace if the predicted interpolation time is of real interest. timeModels in the workspace will be chosen if available.

It is not necessary to load the data set, this happens automatically in predictTime if timeModels if the object is not already existing in the workspace.

Usage

data(timeModels)

Author(s)

Jon Olav Skoien

References

http://www.intamap.org/

[Package intamap version 1.2-08 Index]